requirements.txt
blinker==1.5
certifi==2022.12.7
click==8.1.3
cloudinary==1.30.0
colorama==0.4.6
dnspython==2.2.1
Faker==15.3.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
idna==3.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
MongoFrames==1.3.6
msgpack==1.0.4
pep8==1.7.1
platformdirs==2.5.2
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
rsa==4.9
six==1.16.0
tomli==2.0.1
tomlkit==0.11.5
types-cryptography==3.3.23.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
